refactor file structure, add EditorPlayer
