+++ src/lib/components/player/EditorPlayer.svelte
@@ -1,0 +1,1 @@
+<div />
