+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import MediaPool from "$lib/components/MediaPool.svelte";
-  import Player from "$lib/components/Player.svelte";
+  import Player from "$lib/components/player/Player.svelte";
   import ResizeStalk from "$lib/components/util/ResizeStalk.svelte";
   import Timeline from "$lib/components/Timeline.svelte";
   import { studio, timeline } from "$lib/stores";
